+++ src/Exceptions/InvalidConfigurationException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace MessengerBot\Exceptions;
+
+use RuntimeException;
+
+final class InvalidConfigurationException extends RuntimeException {}